+++ docs/CHANGELOG.md
@@ -16,6 +16,7 @@
 - feat: support skip for pytest
 - feat: print request and response details in DEBUG level when running API cases
 - fix: support None/dict/list format when printing sql response
+- fix: omit pseudo header names for HTTP/1, e.g. :authority, :method, :path, :schema
 
 ## v4.1.3 (2022-06-14)